// backend/src/common/enums/roles-estados.enum.ts
export enum EstadoUsuario {
  ACTIVO = 'ACTIVO',
  BAJA = 'BAJA',
}

export enum RolUsuario {
  MEDICO = 'MEDICO',
  PACIENTE = 'PACIENTE',
  ADMINISTRADOR = 'ADMINISTRADOR',
}

export enum EstadoReserva {
  ACTIVO = 'ACTIVO',
  ATENDIDO = 'ATENDIDO',
  AUSENTE = 'AUSENTE',
  CANCELADO = 'CANCELADO',
}